final version of an app

+++ main.py
@@ -1,6 +1,6 @@
+import bcrypt
 import kivy
-import os
-import sys
+
 import pyperclip
 
 kivy.require('2.3.0')
@@ -9,9 +9,9 @@ from kivy.lang import Builder
 from kivy.config import Config
 from kivy.core.window import Window
 from kivy.uix.screenmanager import ScreenManager, Screen
-from storage_db import add_data, delete_data
-from signup_db import signup
-from hashing import hash_masterpwd, checkpwd
+from storage_db import add_data, delete_data, app_exists, get_password, get_key
+from signup_db import signup, get_masterhash, email_exists
+from hashing import hash_masterpwd
 from encyption import encrypt, decrypt, initialize_cipher, generate_key
 
 # Set configuration
@@ -34,29 +34,37 @@ class LoginPage(Screen):
     def login(self):
         email = self.ids.email.text
         pwd = self.ids.password.text
-        if email.strip() == "" or pwd.strip() == "":
-            print("Please, enter required data to proceed.")
+        if not email or not pwd:  # Check if either email or password is empty
+            self.manager.current = 'wronglog'
             return
-        else:
-            try:
-                stored_mpwd =
-            except ValueError as ve:
-                print("ValueError occurred:", ve)
-            except TypeError as te:
-                print("TypeError occurred:", te)
-            except Exception as e:
-                print("An unexpected error occurred:", e)
+        try:
+            stored_password_hash = get_masterhash(email)
+            if stored_password_hash is not None and bcrypt.checkpw(pwd.encode(), stored_password_hash.encode()):
+                print("Logged in successfully")
+                self.manager.current = 'main'
+                self.ids.email.text = ''
+                self.ids.password.text = ''
+            else:
+                print("Wrong email or password")
+                self.manager.current = 'wronglog'
+        except Exception as e:
+            print("An unexpected error occurred:", e)
 
 
 class RegisterPage(Screen):
     def add_to_signup_db(self):
         email = self.ids.email.text
         masterpwd = self.ids.masterpwd.text
-        hashed_masterpwd, salt = hash_masterpwd(masterpwd)
-        signup(email, hashed_masterpwd, salt)
+        if email_exists(email):
+            self.manager.current = 'wrongreg'
+
+        else:
+            hashed_masterpwd, salt = hash_masterpwd(masterpwd)
+            signup(email, hashed_masterpwd, salt)
+            self.manager.current = 'reggood'
 
 
-class RegfinishPage(Screen):
+class RegisteredGoodPage(Screen):
     pass
 
 
@@ -69,7 +77,6 @@ class AddPage(Screen):
         app_name = self.ids.app_name.text
         email = self.ids.email.text
         pwd = self.ids.password.text
-
         if email.strip() == "" or pwd.strip() == "":
             print("Please, enter required data to proceed.")
             return
@@ -78,7 +85,12 @@ class AddPage(Screen):
                 key = generate_key()
                 cipher = initialize_cipher(key)
                 encrypted_pwd = encrypt(cipher, pwd)
-                add_data(app_name, email, encrypted_pwd)
+                add_data(app_name, email, encrypted_pwd, key)
+                self.manager.current = 'addgood'
+                self.ids.app_name.text = ''
+                self.ids.email.text = ''
+                self.ids.password.text = ''
+
                 print("Password stored securely.")
             except ValueError as ve:
                 print("ValueError occurred:", ve)
@@ -87,24 +99,70 @@ class AddPage(Screen):
             except Exception as e:
                 print("An unexpected error occurred:", e)
 
+
 class DeletePage(Screen):
     def delete_data_from_database(self):
         # Get the app_name from TextInput widget
         app_name = self.ids.app_name.text
+        try:
+            if app_exists(app_name):
+                delete_data(app_name)
+                self.manager.current = 'deletegood'
+            else:
+                self.manager.current = 'deletewrong'
+        except Exception as e:
+            print("An unexpected error occurred:", e)
 
-        # Call delete_data function from the database module
-        delete_data(app_name)
-
-
-class ShowPage(Screen):
+class FetchPage(Screen):
     def fetch_from_db(self):
-        email = self.ids.email.text
-        password = self.ids.password.text
         app_name = self.ids.app_name.text
-        add_data(email, password, app_name)
+
+        if not app_name:
+            self.manager.current = 'fetchbad'
+            return
+        try:
+            key = get_key(app_name)
+            cipher = initialize_cipher(key)
+            encrypted_password = get_password(app_name)
+
+            if encrypted_password:
+                decrypted_password = decrypt(cipher, encrypted_password)
+                if decrypted_password:
+                    pyperclip.copy(decrypted_password)
+                    self.manager.current = 'fetchgood'
+                else:
+                    self.manager.current = 'fetchbad'
+            else:
+                self.manager.current = 'fetchbad'
+        except Exception as e:
+            print("An unexpected error occurred:", e)
+            self.manager.current = 'fetchbad'
+
+class WrongLogPage(Screen):
+    pass
 
 
-class SearchPage(Screen):
+class WrongRegPage(Screen):
+    pass
+
+
+class DeleteGoodPage(Screen):
+    pass
+
+
+class AddDataGoodPage(Screen):
+    pass
+
+
+class DeleteWrongPage(Screen):
+    pass
+
+
+class FetchGoodPage(Screen):
+    pass
+
+
+class FetchBadPage(Screen):
     pass
